requirements.txt
fastapi
uvicorn
crawl4ai
pymupdf
openai
pydantic
python-dotenv
tenacity
json_repair
curl_cffi
httpx
beautifulsoup4
hypercorn
matplotlib
pandas
numpy
mistral-common[sentencepiece]>=0.2.0
mistralai>=1.0.0
replicate>=0.25.0
asyncpg>=0.29.0
pytest>=7.0.0
pytest-asyncio>=0.23.0
arize-phoenix-otel>=0.5.0
openinference-instrumentation-openai>=0.1.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
